+++ addon/rules/after-initial-render.js
@@ -1,0 +1,5 @@
+export default function afterInitialRender({ firstTime, use }) {
+  if (!firstTime) {
+    return use;
+  }
+}