+++ addon/rules/always.js
@@ -1,0 +1,3 @@
+export default function always({ use }) {
+  return use;
+}